+++ packages/block-editor/src/store/defaults.js
@@ -29,6 +29,7 @@ export const PREFERENCES_DEFAULTS = {
  * @property {Array}         __experimentalBlockPatterns            Array of objects representing the block patterns
  * @property {Array}         __experimentalBlockPatternCategories   Array of objects representing the block pattern categories
  * @property {boolean}       __experimentalGenerateAnchors          Enable/Disable auto anchor generation for Heading blocks
+ * @property {boolean}       __experimentalCanLockBlocks            Whether the user can manage Block Lock state
  * @property {boolean}       __unstableGalleryWithImageBlocks       Whether the user has enabled the refactored gallery block which uses InnerBlocks
  */
 export const SETTINGS_DEFAULTS = {
@@ -158,6 +159,7 @@ export const SETTINGS_DEFAULTS = {
 	__experimentalBlockPatternCategories: [],
 	__experimentalSpotlightEntityBlocks: [],
 	__experimentalGenerateAnchors: false,
+	__experimentalCanLockBlocks: true,
 	__unstableGalleryWithImageBlocks: false,
 	// gradients setting is not used anymore now defaults are passed from theme.json on the server and core has its own defaults.
 	// The setting is only kept for backward compatibility purposes.