+++ packages/block-library/src/quote/index.js
@@ -12,12 +12,13 @@ import edit from './edit';
 import metadata from './block.json';
 import save from './save';
 import transforms from './transforms';
+import settingsV2 from './v2';
 
 const { name } = metadata;
 
 export { metadata, name };
 
-export const settings = {
+export const settingsV1 = {
 	icon,
 	example: {
 		attributes: {
@@ -51,3 +52,7 @@ export const settings = {
 	},
 	deprecated,
 };
+
+export const settings = window?.__experimentalEnableQuoteBlockV2
+	? settingsV2
+	: settingsV1;